requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
amqp==2.6.0               # via -r requirements/base.txt, kombu
astroid==1.6.6            # via pylint
backports.functools-lru-cache==1.6.1  # via astroid, isort, pylint
billiard==3.6.3.0         # via -r requirements/base.txt, celery
celery==4.4.6             # via -r requirements/base.txt, eox-core
configparser==4.0.2       # via -r requirements/base.txt, importlib-metadata, pylint
contextlib2==0.6.0.post1  # via -r requirements/base.txt, importlib-metadata, zipp
coverage==5.1             # via -r requirements/test.in
django-crum==0.7.6        # via -r requirements/test.in
django-fake-model==0.1.4  # via -r requirements/test.in
django-filter==1.1.0      # via -c requirements/constraints.txt, -r requirements/test.in
django==1.11.29           # via -c requirements/constraints.txt, -r requirements/base.txt, -r requirements/test.in, django-crum, edx-opaque-keys, eox-core
djangorestframework==3.9.4  # via -r requirements/base.txt, eox-core
edx-opaque-keys[django]==0.4.4  # via -c requirements/constraints.txt, -r requirements/base.txt, -r requirements/test.in, eox-core
enum34==1.1.10            # via astroid
eox-core==2.13.0          # via -r requirements/base.txt
future==0.18.2            # via -r requirements/base.txt, celery
futures==3.3.0 ; python_version == "2.7"  # via -c requirements/constraints.txt, isort
importlib-metadata==1.6.1  # via -r requirements/base.txt, kombu
isort==4.3.21             # via pylint
kombu==4.6.11             # via -r requirements/base.txt, celery
lazy-object-proxy==1.5.0  # via astroid
mccabe==0.6.1             # via pylint
pathlib2==2.3.5           # via -r requirements/base.txt, importlib-metadata
pbr==5.4.5                # via -r requirements/base.txt, stevedore
pycodestyle==2.6.0        # via -r requirements/test.in
pylint==1.9.5             # via -c requirements/constraints.txt, -r requirements/test.in
pymongo==3.10.1           # via -r requirements/base.txt, edx-opaque-keys
pytz==2020.1              # via -r requirements/base.txt, celery, django
scandir==1.10.0           # via -r requirements/base.txt, pathlib2
singledispatch==3.4.0.3   # via astroid, pylint
six==1.15.0               # via -r requirements/base.txt, astroid, edx-opaque-keys, pathlib2, pylint, singledispatch, stevedore
stevedore==1.32.0         # via -r requirements/base.txt, edx-opaque-keys
vine==1.3.0               # via -r requirements/base.txt, amqp, celery
wrapt==1.12.1             # via astroid
zipp==1.2.0               # via -r requirements/base.txt, importlib-metadata
